Fix circular dependencies
The rollup build detected multiple circular dependencies. These
are now refactored.

Also adapted .gitignore and .npmignore to include the visualizer
output.

Slightly adjusts the README.
