+++ README.md
@@ -34,8 +34,8 @@ Linting, formatting and import sorting should work automatically.
 
 ### Scripts
 
+- Build and watch jsonforms-editor library with `npm run watch`
 - Start the app with `npm start`
-- Watch changes on the jsonforms-editor library with `npm run watch` (concurrently with previous command)
 - Run unit tests with `npm run test`
 - Run UI tests with `npm run cypress:open`
