+++ jsonforms-editor/.gitignore
@@ -1,1 +1,2 @@
 dist/
+stats.html
