+++ jsonforms-editor/.npmignore
@@ -1,1 +1,2 @@
 node_modules/
+stats.html
